update-dashboard-issue: add date updated
This commit adds the date last updated to the dashboard issue. By doing so, it will be more obvious to translators and maintainers that the comment is been updated regularly.

It may also be useful at a future time to link to this in the main README in tldr-pages/tldr.